Immutable persistence and fixes
[jacodb-api-jvm] Clarify JcDatabase.setImmutable()

As of this commit, JcDatabase.setImmutable() doesn't try to implicitly load
earlier saved database dump. Instead, it tries to do that in JcDatabase.restore().

[jacodb-storage] Represent instance id lists as wrappers atop of ByteBuffers

On loading dump from a file, lists of instance ids do wrap mapped ByteBuffers.
Any delta coding is no longer used, so mapping doesn't require any reading of
mapped ByteBuffers.

InputStream & OutputStream extensions are cleaned up.

[jacodb-storage] Fix readUnsignedOrderedList()

[jacodb-storage] Fix filling instance id lists for AttributesImmutable

[jacodb-storage] Add and use LongRangeList

For a long range, do not instantiate a list of all longs of the progression,
but use special LongRangeList. This reduces time spent on list population
and memory used as well.

[jacodb-storage] Represent instance ids of AttributesImmutable as List, not array

This simple change reduces unnecessary memory traffic.
Some code clean-up is being done.

[jacodb-storage] On loading AttributesImmutable, avoid checking if values are sorted

On dumping an AttributesImmutable instance, it's already known if its values are naturally
ordered or not, so this information can be stored as well. So on loading, the check
if the values are sorted can be skipped.

[jacodb-benchmarks] Add JcdbRAMImmutableIdeaBackgroundBenchmarks

[jacodb-storage] Fix ensuring capacity in case if it's >= 2^30

[jacodb-api-jvm] Add JcByteCodeLocation.fileSystemIdHash

Unlike JcByteCodeLocation.currentHash, JcByteCodeLocation.fileSystemIdHash
is being calculated once. It is used to calculate id of JcDatabase.
This allows

[jacodb-storage] Implement RAMEntityRelationshipStorage.load(databaseId)

[jacodb-api-jvm, jacodb-storage] Add ability to load immutable code model by id prior to indexing

On attempt to invoke JcDatabase.setImmutable(), try at first to force persistence to
load the model by id. If it fails, do as always: wait for background jobs that build
the model, and then convert it to immutable one.
API additions:
- JcDatabasePersistence.tryLoad(databaseId) added; it has non-trivial implementation
in ErsPersistenceImpl only;
- DumpableLoadableEntityRelationshipStorage.load(databaseId) added.
Any implementation of DumpableLoadableEntityRelationshipStorage used with ErsPersistenceImpl
can automatically provide functionality of dumping/loading immutable code models.

[jacodb-api-jvm, jacodb-core, jacodb-storage] Use JcDatabase.id when setting the database to immutable

For a JcDatabase, use its id as a parameter for setImmutable() method. Provide
RamErsSettings with an optional path to dump the storage. If this path is set,
RAMEntityRelationshipStorage would automatically dump itself in the asImmutable()
method, if corresponding files doesn't exist. If it does, the storage would load
itself from the dump file.

[jacodb-api-jvm, jacodb-core] Add JcDatabase.id

For a JcDatabase, id is xor of all hashes of its locations.

[jacodb-api-jvm] For JarLocationImpl, use SHA-256 of its content as identity

[jacodb-api-jvm] Define JcByteCodeLocation identify as currentHash: BigInteger

The currentHash property returns SHA-256 hash of location as it was
initially defined for various types of locations.

[jacodb-storage] Represent attributes' offsets & lengths with LongBuffer

Get rid of OffsetsAndLensCollection interface and all its implementations.
As of this commit, attributes' offsets & lengths are stored and loaded
as fixed-length longs. ByteBuffer is used for that, so it can be loaded as
MappedByteBuffer, thus completely freeing heap memory from offsets & lengths info.

[jacodb-storage] Fix compilation with JDK 1.8

In Java 1.8, ByteBuffer methods return instances of base Buffer class.

[jacodb-storage] Optionally use memory-mapped files for loading attribute values

[jacodb-api-storage, jacodb-storage] Add ability to dump/load immutable RAMEntityRelationshipStorage

As of this commit, RAMEntityRelationshipStorage inherits dedicated interface
`DumpableLoadableEntityRelationshipStorage`. Only immutable RAMEntityRelationshipStorage
can be dumped. Any RAMEntityRelationshipStorage can be loaded as immutable storage.

[jacodb-storage] Greenify

[jacodb-storage] For accessing attribute values, use ByteBuffer instead of ByteArray

Minor refactoring is being done: for AttributesImmutable, Properties, Links &
RAMDataContainerImmutable, all extension functions moved to separate files.

[jacodb-storage] Make SparseBitSet Iterable<Long>

[jacodb-storage] Optimize EntityRelationshipStorage.asImmutable

[jacodb-benchmarks] Fix compilation

[jacodb-api-storage] Clean up ERS API: rename .asReadonly to .asImmutable

[jacodb-cli] Make jacodb-cli dependent on jacodb-storage

[jacodb-api-storage] Clean up API: get rid of PluggableKeyValueStorage.readonly

[jacodb-storage] Add ability to create immutable RAMEntityRelationshipStorage from arbitrary EntityRelationshipStorage

EntityRelationshipStorageBase added to define default implementation of EntityRelationshipStorage.asReadonly.

[jacodb-api-storage] Add Transaction.getTypes()

Added a method returning all available pairs of entity types and their ids.

[jacodb-storage] Compress representation of offsets/lengths of attributes' values

Further compression of offsets added by representation of them with short values.

[jacodb-storage] Compress representation of offsets/lengths of attributes' values

Added OffsetsAndLensCollection interface and two implementations:
LongArrayOffsetsAndLensCollection & IntArraySameLenOffsetsAndLensCollection.
The first one just wraps LongArray, whereas the second wraps an
array of offsets of values all having the same length.